chore: separate route and service
