+++ packages/element-api/src/express/routes/index.js
@@ -3,9 +3,7 @@ const express = require('express');
 const router = express.Router();
 
 router.use('/version', require('./version'));
-
 router.use('/sidetree', require('./sidetree'));
-
-router.use('/sidetree-lite', require('./sidetree-lite'));
+router.use('/sidetree-v2', require('./sidetree-v2'));
 
 module.exports = router;
